refactor: hostCompliedFile as a independent methods
